chore: use given/then/when pattern for tests
